Added support for more CUSTOM backends recognition Fixed excessive number of hyperlinks corrupting large reports Added all components of Snapshot report

+++ ReportObjects/ActivityFlow/ActivityFlow.cs
@@ -39,7 +39,7 @@ namespace AppDynamics.Dexter.DataObjects
         public double ErrorsPercentage { get; set; }
 
         public string MetricLink { get; set; }
-        public List<int> MetricsIDs { get; set; }
+        public List<long> MetricsIDs { get; set; }
 
         public ActivityFlow Clone()
         {
